+++ src/EndpointConfig.php
@@ -11,8 +11,8 @@ use Nette\PhpGenerator\ClassType;
 use Spawnia\Sailor\Codegen\DirectoryFinder;
 use Spawnia\Sailor\Codegen\Finder;
 use Spawnia\Sailor\Error\Error;
-use Spawnia\Sailor\Event\EndRequest;
-use Spawnia\Sailor\Event\StartRequest;
+use Spawnia\Sailor\Events\ReceiveResponse;
+use Spawnia\Sailor\Events\StartRequest;
 use Spawnia\Sailor\Type\BooleanTypeConfig;
 use Spawnia\Sailor\Type\EnumTypeConfig;
 use Spawnia\Sailor\Type\FloatTypeConfig;
@@ -52,11 +52,12 @@ abstract class EndpointConfig
     abstract public function schemaPath(): string;
 
     /**
-     * Instantiate a class to find GraphQL documents.
+     * Will be called with events that happen during the execution lifecycle.
+     *
+     * @param StartRequest|ReceiveResponse $event
      */
-    public function finder(): Finder
+    public function handleEvent(object $event): void
     {
-        return new DirectoryFinder($this->searchPath());
     }
 
     /**
@@ -115,6 +116,14 @@ abstract class EndpointConfig
         return [];
     }
 
+    /**
+     * Instantiate a class to find GraphQL documents.
+     */
+    public function finder(): Finder
+    {
+        return new DirectoryFinder($this->searchPath());
+    }
+
     public function typesNamespace(): string
     {
         return $this->namespace() . '\\Types';
@@ -128,12 +137,5 @@ abstract class EndpointConfig
     public function typeConvertersNamespace(): string
     {
         return $this->namespace() . '\\TypeConverters';
-    }
-
-    /**
-     * @param StartRequest|EndRequest $event
-     */
-    public function fireEvent($event): void
-    {
     }
 }