+++ src/Operation.php
@@ -5,8 +5,9 @@ namespace Spawnia\Sailor;
 use Mockery;
 use Mockery\MockInterface;
 use Spawnia\Sailor\Convert\TypeConverter;
-use Spawnia\Sailor\Event\EndRequest;
-use Spawnia\Sailor\Event\StartRequest;
+use Spawnia\Sailor\Events\ReceiveResponse;
+use Spawnia\Sailor\Events\StartRequest;
+use stdClass;
 
 /**
  * Subclasses of this class are automatically generated.
@@ -57,7 +58,7 @@ abstract class Operation implements BelongsToEndpoint
             return $mock::execute(...$args);
         }
 
-        $response = self::fetchResponse($args);
+        $response = static::fetchResponse($args);
 
         $child = static::class;
         $parts = explode('\\', $child);
@@ -76,7 +77,27 @@ abstract class Operation implements BelongsToEndpoint
      */
     protected static function fetchResponse(array $args): Response
     {
-        $variables = new \stdClass();
+        $endpointConfig = Configuration::endpoint(static::config(), static::endpoint());
+
+        $document = static::document();
+        $variables = static::variables($args);
+
+        $endpointConfig->handleEvent(new StartRequest($document, $variables));
+
+        $client = self::$clients[static::class] ?? $endpointConfig->makeClient();
+        $response = $client->request($document, $variables);
+
+        $endpointConfig->handleEvent(new ReceiveResponse($response));
+
+        return $response;
+    }
+
+    /**
+     * @param array<int, mixed> $args
+     */
+    protected static function variables(array $args): stdClass
+    {
+        $variables = new stdClass();
         $arguments = static::converters();
         foreach ($args as $index => $arg) {
             if (ObjectLike::UNDEFINED === $arg) {
@@ -87,18 +108,7 @@ abstract class Operation implements BelongsToEndpoint
             $variables->{$name} = $typeConverter->toGraphQL($arg);
         }
 
-        $endpointConfig = Configuration::endpoint(static::config(), static::endpoint());
-
-        $client = self::$clients[static::class]
-            ?? $endpointConfig->makeClient();
-
-        $document = static::document();
-
-        $endpointConfig->fireEvent(new StartRequest($document, $variables));
-        $response = $client->request($document, $variables);
-        $endpointConfig->fireEvent(new EndRequest($response));
-
-        return $response;
+        return $variables;
     }
 
     /**
@@ -106,9 +116,8 @@ abstract class Operation implements BelongsToEndpoint
      */
     public static function mock(): MockInterface
     {
-        // @phpstan-ignore-next-line The type of MockInterface matches up, I promise
-        return self::$mocks[static::class]
-            ??= Mockery::mock(static::class);
+        // @phpstan-ignore-next-line I solemnly swear the type of MockInterface matches
+        return self::$mocks[static::class] ??= Mockery::mock(static::class);
     }
 
     public static function clearMocks(): void